データ詳細から入力処理を作成する。

+++ database/migrations/2020_07_10_092328_create_eating_management.php
@@ -17,10 +17,10 @@ class CreateEatingManagement extends Migration
             $table->increments('id');
             $table->date('date');
             $table->text('item');
-            $table->integer('protein');
-            $table->integer('liqid');
-            $table->integer('carbo');
-            $table->integer('calorie');
+            $table->double('protein');
+            $table->double('liqid');
+            $table->double('carbo');
+            $table->double('calorie');
             $table->timestamps();
             $table->engine = 'InnoDB';
             $table->charset = 'utf8mb4';
